move function to get API key to data_validation
 * no need to import it from end_of_run_workflow()
